Change BIGINT cast to INTEGER cast in q54
It was cast to INT in original TPCDS query

+++ presto-product-tests/src/main/resources/sql-tests/testcases/tpcds/q54.sql
@@ -61,7 +61,7 @@ UNION ALL       SELECT
    GROUP BY "c_customer_sk"
 ) 
 , segments AS (
-   SELECT CAST(("revenue" / 50) AS BIGINT) "segment"
+   SELECT CAST(("revenue" / 50) AS INTEGER) "segment"
    FROM
      my_revenue
 )
